Fake deterministic encryption with age
Do no re-encrypt files if neither plaintex or the recipient file have
changed. This gives us the equivalent behaviour to SIV encryption.

Update the doc to target age in the primary usage example, keeping the
SIV relevant parts.

Add tests for age flow.
