+++ Dockerfile
@@ -6,4 +6,4 @@ ENV GOPATH=/go CGO_ENABLED=0
 COPY . /go/src/github.com/uw-labs/strongbox
 WORKDIR /go/src/github.com/uw-labs/strongbox
 
-ENTRYPOINT ["/bin/sh", "/go/src/github.com/uw-labs/strongbox/integration_tests/run.sh"]
+ENTRYPOINT ["/bin/sh", "/go/src/github.com/uw-labs/strongbox/run_tests"]
